requirements.txt
spacy-nightly==2.0.0a
numpy
tensorflow
